+++ adminer/include/editing.inc.php
@@ -268,7 +268,7 @@ function process_in(string $val): string {
 * @param FieldType $field
 */
 function process_type(array $field, string $collate = "COLLATE"): string {
-	return " $field[type]"
+	return " " . (is_user_type($field["type"]) ? idf_escape($field["type"]) : $field["type"])
 		. process_length($field["length"])
 		. (preg_match(number_type(), $field["type"]) && in_array($field["unsigned"], driver()->unsigned) ? " $field[unsigned]" : "")
 		. (preg_match('~' . text_type() . '~', $field["type"]) && $field["collation"] ? " $collate " . (JUSH == "mssql" ? $field["collation"] : q($field["collation"])) : "")